requirements.txt
rich
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv
ddgs